feat(forge): restore full 1.20.1 port
- Port all eight legacy modules with original resources and identities
- Add migration, GUI, recipe, networking, and gameplay regression coverage
- Keep release blocked on the remaining upward Chute GameTest

+++ ports/forge-1.20.1/src/main/java/buildcraft/neo/forge1201/factory/VerticalTankPolicy.java
@@ -27,6 +27,6 @@ public final class VerticalTankPolicy {
             return 0;
         }
         int clampedAmount = Math.min(amount, capacity);
-        return Math.min(15, (clampedAmount * 14) / capacity + 1);
+        return Math.min(15, (int) (((long) clampedAmount * 14L) / capacity) + 1);
     }
 }
